Remove duplication in the TryFrom<Request> impl blocks
We can collapse all those impls into a single one that is bounded
by the correct trait bounds for the individual items.
